+++ ReSwift/CoreTypes/Action.swift
@@ -3,7 +3,7 @@
 //  ReSwift
 //
 //  Created by Benjamin Encz on 12/14/15.
-//  Copyright © 2015 Benjamin Encz. All rights reserved.
+//  Copyright © 2015 ReSwift Community. All rights reserved.
 //
 
 /// All actions that want to be able to be dispatched to a store need to conform to this protocol